用户模块基础

+++ src/main/java/org/fxi/quick/common/entity/BaseEntity.java
@@ -5,7 +5,6 @@ import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableLogic;
-import com.baomidou.mybatisplus.annotation.Version;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
 import lombok.Data;
@@ -50,5 +49,5 @@ public class BaseEntity {
     private LocalDateTime deleteTime;
 
     @TableLogic
-    private Integer delFlag;
+    private Short delFlag;
 }
